Moved license check out of db to its own utils file, added newlines to break apart license check function, and removed an unnecessary boolean coercion
